requirements.txt
playwright
beautifulsoup4
requests
python-dotenv
time
datetime
lxml